add renderTasks.js
